Attempt to fix a flushing problem
The point is that when sessions are flushed at the same time as customer
flushing is initiated in the background, there's a chance that a
customer hasn't yet been written to the database by the time the session
record is written. This was causing the FK constraint to fire off. To
fix this, I needed to split flushing into several stages and allow the
session observer to collect user IDs from its session updates, and then
initiate another customer flushing with these IDs. This may seem
sub-optimal, but since WBUpdateIterator isn't emitting update records
that are either already locked or empty, the only possible performance
hit we get is having two parallel transactions on the server.

What still worries me is that there is a chance that a session update
could be written to the DB while the transaction in which the customer
has been processed is not yet over. So, the current fix is still not
perfect.